backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Configuration
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Utilities
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
